# Conflicts:
#	admin/videos/updateVideo.php

+++ admin/videos/updateVideo.php
@@ -3,7 +3,11 @@ include "../../connect.php";
 include "../sessionAdmin.php";
 $video_id = $_GET['videoID'];
 //check if record exist
-
+$editQuestionID=0;
+if(isset($_GET['editQuestionID']))
+{
+	$editQuestionID = $_GET['editQuestionID'];
+}
 
 ?>
 <!DOCTYPE html>
@@ -14,7 +18,7 @@ $video_id = $_GET['videoID'];
     <meta name="description" content="">
     <meta name="keywords" content="">
     <meta charset="UTF-8">
-    <title>User View</title>
+    <title>Video Details</title>
 
     <link href="https://fonts.googleapis.com/css?family=Poppins:100,200,400,300,500,600,700" rel="stylesheet">
     <!--
@@ -34,7 +38,7 @@ $video_id = $_GET['videoID'];
 <header id="header">
     <div class="header-top">
         <div class="container">
-           
+
         </div>
     </div>
     <div class="container main-menu">
@@ -56,8 +60,8 @@ $video_id = $_GET['videoID'];
 				              <li><a href="../changePassword/changePassword.php">Change Password</a></li>
 							  <li><a href="../logoutSessionAdmin.php">Logout</a></li>
 				            </ul>
-				          </li>	
-				        
+				          </li>
+
 				        </ul>
 				      </nav><!-- #nav-menu-container -->
 
@@ -87,38 +91,75 @@ if(mysqli_num_rows($result1)>0)
     }
 }
 ?>
-<section class="destinations-area pt-20">z
-<form action="updateVideoDetails.php?id=<?php echo $video_id; ?>" method=post enctype = "multipart/form-data">
+<section class="destinations-area pt-20">
+    <?php
+    if(isset($_GET['editVideo']))
+    {
+?>
+    <form action="updateVideoDetails.php?id=<?php echo $video_id; ?>" method=post enctype = "multipart/form-data">
     <h1 class="pb-10 d-flex justify-content-center"><b>  </b></h1>
     <div class="container">
 	<input type="text" class="form-control" style="width:100%"  name="nameTxt" placeholder="Name " onfocus="this.placeholder = ''" onblur="this.placeholder = 'Name '" value= "<?php echo $video_title;?>">
         <br><div class="row">
             <div class="map-wrap" style="width:100%; height: 445px; box-shadow: 0px 10px 30px 0px rgba(60, 64, 143, 0.3);">
                 <embed style="width:100%; height: 400px; play="true" loop="false" menu="true" src="<?php echo $path; ?>">
-				<b>Video File</b>
-				<input type="file" name="videoFile" >
-				<input type="hidden" name="videoFile_tmp" value = "<?php echo $path; ?>" >
-				<b>
-				Thumbnail
-				</b>
-				<input type="file" name="thumbnailFile" >
-				<input type="hidden" name="thumbnailFile_tmp" value = "<?php echo $path; ?>" >
-				<br><br>
-            </div>
+    <b>Video File</b>
+    <input type="file" name="videoFile" >
+    <input type="hidden" name="videoFile_tmp" value = "<?php echo $path; ?>" >
+    <b>
+        Thumbnail
+    </b>
+    <input type="file" name="thumbnailFile" >
+    <input type="hidden" name="thumbnailFile_tmp" value = "<?php echo $path; ?>" >
+    <br><br>
+    </div>
+    </div>
+    <div class="row">
+        <div class="col-lg-2">
         </div>
-
-        <div class="row d-flex justify-content-center">
-            <div class=" pb-20 pt-20 col-lg-8 text-center">
-               <textarea class="form-control" rows=7 name="descriptionTxt" placeholder="Description" onfocus="this.placeholder = ''" onblur="this.placeholder = 'Description'" required>
-				<?php echo $desc; ?>
-				</textarea>
-            </div>
+        <div class="pb-20 pt-20 col-lg-8">
+            <textarea class="form-control" rows=7 name="descriptionTxt" placeholder="Description" onfocus="this.placeholder = ''" onblur="this.placeholder = 'Description'" required>
+                <?php echo $desc; ?>
+            </textarea>
         </div>
+    </div>
 
     </div><div class="container">
-		 <button id="btn_submit" name="save_video" class="primary-btn text-uppercase text-center" style="float:right">SAVE VIDEO</button>
-		 </div>
-	</form>
+            <a href="updateVIdeo.php?videoID=<?php echo $video_id; ?>" ><button class="primary-btn text-uppercase text-center" style="float:right; background-color:gray;">CANCEL</button>
+        <button id="btn_submit" name="save_video" class="primary-btn text-uppercase text-center" style="float:right; margin-right: 10px">SAVE VIDEO</button>
+
+    </div>
+    </form>
+<?php
+    }
+    else
+    {
+?>
+            <h1 class="pb-10 d-flex justify-content-center">
+            <b> <?php echo $video_title;?></b>
+            </h1>
+
+        <div class="container">
+            <div class="row">
+                <div class="map-wrap" style="width:100%; height: 445px; box-shadow: 0px 10px 30px 0px rgba(60, 64, 143, 0.3);">
+                    <embed style="width:100%; height: 400px; play="true" loop="false" menu="true" src="<?php echo $path?>">
+                </div>
+            </div>
+            <div class="row d-flex justify-content-center">
+                <div class=" pb-20 col-lg-8 text-center">
+                    <p class="pt-20 font-weight-bold"><?php echo $desc; ?></p>
+                </div>
+            </div>
+            <div class="row">
+                <div class="col-lg-12">
+                <a href="updateVideo.php?videoID=<?php echo $video_id; ?>&editVideo=true" class="genric-btn info radius" style="float:right; background-color: #f8b600" >UPDATE VIDEO DETIALS</a>
+            </div>
+                </div>
+        </div>
+   <?php
+   }
+    ?>
+
 </section>
 <br><br><br>
 
@@ -127,10 +168,10 @@ if(mysqli_num_rows($result1)>0)
 <div class="container">
 <h1 style="text-align:center">Exam</h1>
 <a href="#" class="genric-btn info radius" style="float:right;"  data-toggle="modal" data-target="#addQuestionModal" >Add Question</a>
-</div>	
-<br>		
+</div>
+<br>
 <br><br>
-    <form class="form-wrap" action="checkUserAnswer.php" method="POST">
+    <form class="form-wrap" action="updateQuestion.php?videoID=<?php echo $video_id; ?>&questionID=<?php echo $editQuestionID; ?>" method="POST">
         <div class="container" style="width:90%; box-shadow: 0px 10px 10px 0px rgba(60, 64, 143, 0.3);">
             <div class="row d-flex justify-content-center">
                 <div class="menu-content col-lg-8">
@@ -149,13 +190,40 @@ if(mysqli_num_rows($result1)>0)
                         $evalId = $row2['id'];
                         ?>
                         <div class="col-lg-12">
-						<button type="button" style="float:right; margin:5px; padding:0" class="close" data-toggle="modal" data-target="#deleteQuestionModal" onclick="deleteQuestionID(<?php echo $row2['id']; ?>)"><i class="fa fa-times" aria-hidden="true"></i></button>
-                            <div class="single-price pt-5 pb-0">
-							
+						<?php
+						if($editQuestionID == $evalId)
+						{
+							?>
+							<?php
+						}
+						else
+						{
+							?>
+							<button type="button" style="float:right; margin:5px; padding:0" class="close" data-toggle="modal" data-target="#deleteQuestionModal" onclick="deleteQuestionID(<?php echo $row2['id']; ?>)"><i class="fa fa-times" aria-hidden="true"></i></button>
+							<i style="float:right; margin:10px; padding:0; font-size: 1.5em" class="fa fa-edit close" aria-hidden="true" onclick="editQuestionID(<?php echo $row2['id']; ?>)" > </i>
+							<?php
+						}
+
+
+						?>
+
+                            <div class="single-price pt-5 pb-0" style=" <?php if($editQuestionID == $evalId) echo "background-color:white"; ?>">
                                 <h4 class="text-uppercase">
-								<input type="text" class="form-control" name="question_<?php echo $row2['id']; ?>" value="<?php echo $row2['question']?>">
+								<?php
+								if($editQuestionID == $evalId)
+								{
+								?>
+								<input type="text" class="form-control" name="questionTxt" value="<?php echo $row2['question']?>">
+								<?php
+								}
+								else
+								{
+									echo $row2['question'];
+								}
+								?>
+
 								</h4>
-                               
+
 								<ul class="price-list">
                                     <?php
                                     $sql3 = "SELECT B.id, B.choice
@@ -163,26 +231,69 @@ if(mysqli_num_rows($result1)>0)
                                         INNER JOIN evaluation_choices B ON A.id = B.evaluationID
                                         WHERE A.id = '$evalId' and B.status = 1";
                                     $result3 = mysqli_query($con, $sql3);
+                                    $choiceCounter=0;
                                     if(mysqli_num_rows($result3)>0)
                                     {
-                                        while($row3 = mysqli_fetch_array($result3)){
-                                            ?>
-                                            <li class="d-flex justify-content-between align-items-center">
-                                                <p><label class="text-uppercase">
-												<input type="text" class="form-control" name="question_<?php echo $row2['id']; ?>_choice_<?php echo $row3['id']; ?>" value="<?php echo $row3['choice'] ?>">
-												</label>
-												<button type="button" style="float:right; margin-left:10px;" class="btn btn-danger" data-toggle="modal" data-target="#deleteChoiceModal" onclick="deleteChoiceID(<?php echo $row3['id'];?>)"><i class="fa fa-trash" aria-hidden="true"></i></button>
-												
-												</p></li>
-											
-                                        <?php
+                                        while($row3 = mysqli_fetch_array($result3))
+                                        {
+                                            if($editQuestionID == $evalId)
+											{
+												?>
+													<li class="d-flex justify-content-between align-items-center">
+														<p>
+															<label class="text-uppercase">
+                                                                <input type="hidden" class="form-control" name="choiceID_<?php echo $choiceCounter; ?>" value="<?php echo $row3['id'] ?>">
+                                                                <input type="text" class="form-control" name="choiceTxt_<?php echo $choiceCounter; ?>" value="<?php echo $row3['choice'] ?>">
+															</label>
+
+														</p>
+													</li>
+												<?php
+											}
+											else
+											{
+												?>
+													<li class="d-flex justify-content-between align-items-center">
+														<p>
+															<label class="text-uppercase">
+																<?php echo $row3['choice'] ?>
+															</label>
+															<button type="button" style="float:right; margin-left:10px;" class="btn btn-danger" data-toggle="modal" data-target="#deleteChoiceModal" onclick="deleteChoiceID(<?php echo $row3['id'];?>)"><i class="fa fa-trash" aria-hidden="true"></i></button>
+														</p>
+													</li>
+												<?php
+											}
+						                       $choiceCounter++;
                                         }
+
+                                        ?>
+                                        <input type="hidden" class="form-control" name="choiceCounter" value="<?php echo $choiceCounter; ?>">
+                                        <?php
                                     }
                                     ?>
                                 </ul>
-								<a href="#" class="genric-btn success radius" style="float:left;"  data-toggle="modal" data-target="#addChoiceModal" onclick="choicesIDChange(<?php echo $row2['id']; ?>)" >Add Choices</a>
+
+								<?php if($editQuestionID == $evalId)
+											{
+												?>
+
+
+								 <button id="btn_submit" name="save_video" class="primary-btn text-uppercase text-center" style="float:right; background-color:gray;">CANCEL</button>
+								 <button id="btn_submit" name="save_video" class="primary-btn text-uppercase text-center" style="float:right; margin-right: 10px;">SAVE</button>
+								<?php
+											}
+								else
+								{
+									?>
+									<a href="#" class="genric-btn success radius" style="float:left;"  data-toggle="modal" data-target="#addChoiceModal" onclick="choicesIDChange(<?php echo $row2['id']; ?>)" >Add Choices</a>
+
+									<?php
+								}
+								?>
 								<br><br><hr>
+
                             </div>
+
                         </div>
 
                     <?php
@@ -190,23 +301,23 @@ if(mysqli_num_rows($result1)>0)
                 }
                 ?>
             </div>
-			
+
         </div>
 		<br>
 		<div class="container">
-		<button id="btn_submit" name="save_exam" class="primary-btn text-uppercase text-center" style="float: right">SAVE EXAM</button>
+
 		</div><br><br>
-            
-		
-                
-        
+
+
+
+
     </form>
 </section>
 
 <div class="modal animated zoomIn" id="addQuestionModal">
 				<div class="modal-dialog" role="document">
 					<div class="modal-content">
-						<form class="form-horizontal" action="addQuestion.php?videoID=<?php echo $video_id; ?>" method=post> 
+						<form class="form-horizontal" action="addQuestion.php?videoID=<?php echo $video_id; ?>" method=post>
 							<div class="modal-header">
 							<h3 style="text-align:center">Add Question</h3>
 								<button type="button" class="close" data-dismiss="modal">×</button>
@@ -215,7 +326,7 @@ if(mysqli_num_rows($result1)>0)
 								<textarea class="form-control" name="questionTxt" placeholder="Question" onfocus="this.placeholder = ''" onblur="this.placeholder = 'Question'" required></textarea>
 								<br>
 								<button type="submit" class="btn btn-primary"  style="width:100%" name="residentSubmit">Save</button>
-								
+
 							</div>
 						</form>
 					</div>
@@ -224,7 +335,7 @@ if(mysqli_num_rows($result1)>0)
 <div class="modal animated zoomIn" id="addChoiceModal">
 				<div class="modal-dialog" role="document">
 					<div class="modal-content">
-						<form class="form-horizontal" action="addChoice.php" method=post id=addChoiceForm > 
+						<form class="form-horizontal" action="addChoice.php" method=post id=addChoiceForm >
 							<div class="modal-header">
 							<h3 style="text-align:center">Add Choices</h3>
 								<button type="button" class="close" data-dismiss="modal">×</button>
@@ -233,23 +344,23 @@ if(mysqli_num_rows($result1)>0)
 								<input type=text class="form-control" name="choiceTxt" placeholder="Choices" onfocus="this.placeholder = ''" onblur="this.placeholder = 'Choices'" required>
 								<br>
 								<button type="submit" class="btn btn-primary"  style="width:100%" name="residentSubmit">Save</button>
-								
+
 							</div>
 						</form>
 					</div>
 				</div>
 			</div>
-			
+
 			<div class="modal animated zoomIn" id="deleteQuestionModal">
 				<div class="modal-dialog" role="document">
 					<div class="modal-content">
-						<form class="form-horizontal" action="deleteQuestion.php" method=post id="deleteQuestionForm"> 
+						<form class="form-horizontal" action="deleteQuestion.php" method=post id="deleteQuestionForm">
 							<div class="modal-header">
 							<h3 style="text-align:center">Delete Question</h3>
 								<button type="button" class="close" data-dismiss="modal">×</button>
 							</div>
 							<div class="modal-body">
-								<p>Are you sure you want to delete this question?</p>	
+								<p>Are you sure you want to delete this question?</p>
 								<br>
 								<button type="submit" class="btn btn-primary" style="width:45%" name="residentSubmit">Yes</button>
 								<button class="btn btn-default" data-dismiss="modal" style="width:45%; float: right">Cancel</button>
@@ -258,17 +369,17 @@ if(mysqli_num_rows($result1)>0)
 					</div>
 				</div>
 			</div>
-			
+
 			<div class="modal animated zoomIn" id="deleteChoiceModal">
 				<div class="modal-dialog" role="document">
 					<div class="modal-content">
-						<form class="form-horizontal" action="deleteChoice.php" method=post id="deleteChoiceForm"> 
+						<form class="form-horizontal" action="deleteChoice.php" method=post id="deleteChoiceForm">
 							<div class="modal-header">
 							<h3 style="text-align:center">Delete Choice</h3>
 								<button type="button" class="close" data-dismiss="modal">×</button>
 							</div>
 							<div class="modal-body">
-								<p>Are you sure you want to delete this choice?</p>	
+								<p>Are you sure you want to delete this choice?</p>
 								<br>
 								<button type="submit" class="btn btn-primary" style="width:45%" name="residentSubmit">Yes</button>
 								<button class="btn btn-default" data-dismiss="modal" style="width:45%; float: right">Cancel</button>
@@ -301,19 +412,25 @@ if(mysqli_num_rows($result1)>0)
     $('#btn_view_exam').click(function(){
         $('#exam_collapse').slideToggle('slow');
     });
-	
+
 	function choicesIDChange(id)
 	{
 		document.getElementById("addChoiceForm").action = "addChoice.php?videoID=<?php echo $video_id; ?>&questionID="+id;
 	}
-	
+
 	function deleteQuestionID(id)
 	{
 		document.getElementById("deleteQuestionForm").action = "deleteQuestion.php?videoID=<?php echo $video_id; ?>&questionID="+id;
 	}
-	
+
 	function deleteChoiceID(id)
 	{
 		document.getElementById("deleteChoiceForm").action = "deleteChoice.php?videoID=<?php echo $video_id; ?>&choiceID="+id;
+	}
+
+	function editQuestionID(id)
+	{
+		window.location.href = 'updateVideo.php?videoID=<?php echo $video_id; ?>&editQuestionID='+id;
+
 	}
 </script>
